+++ src/main/java/com/yjz/notepad/service/IUserService.java
@@ -13,9 +13,10 @@ public interface IUserService {
 
     /**
      * @param user user
-     * @return 成功的状态
+     * @return 插入的主键ID
+     *
      */
-    int addUserByObject(User user);
+    Long addUserByObject(User user);
 
     /**
      * @param id userID
